requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.3
idna==3.4
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
openai==0.27.8
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pvporcupine==2.2.1
PyAudio==0.2.13
pydantic==2.0
pydantic_core==2.0.2
pydub==0.25.1
Pygments==2.15.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
yarl==1.9.2

playsound~=1.3.0